Fixed DL3 creation complaining about mismatch in binning.
It sometimes happened that the IRF files have different bin ranges saved despite using the same input card.
The difference is very small, comparable to the numerical accuracy of double precision.
I saw it in the files processed in CNAF and my suspicion is that it comes from inhomogenuity of machines over which the code was run.
Now if different binning is detected, it checks also if the relative difference exceeds somewhere 10^-15 and only then raises an error.